テンプレート仕様の変更
