+++ assets/package.json
@@ -57,7 +57,6 @@
     "vinyl-named": "^1.1.0",
     "vue-loader": "^16.8.3",
     "vue-style-loader": "^4.1.3",
-    "vue-styleguidist": "^4.72.4",
     "vue-template-compiler": "^2.7.16",
     "webpack": "^5.90.3",
     "webpack-stream": "^7.0.0"
