--- assets/src/_functions/vueAppFactory.ts
@@ -1,9 +1,0 @@
-import { createApp, defineComponent } from "vue";
-import store from "@/_store/index";
-
-export default function vueAppFactory(
-  component: ReturnType<typeof defineComponent>,
-  target: string
-) {
-  createApp(component).use(store).mount(target);
-}
