+++ assets/src/client/_components/common/Footer.vue
