+++ assets/src/client/_components/icons/Close.vue
