+++ assets/src/client/_components/video/cards/VideoCard.vue
