+++ assets/src/client/_functions/isMobileDevice.ts
