+++ assets/src/client/_functions/smoothScroll.ts
@@ -1,0 +1,36 @@
+// import store from "@/client/_store/index";
+
+/**
+ * スムーズスクロールを実現する関数です。
+ * @param target - スクロール先の要素のセレクター
+ * @param duration - スクロールのアニメーション時間（ミリ秒）。デフォルト値は500です。
+ */
+export default function smoothScroll(
+  target: string,
+  duration: number = 500
+): void {
+  const targetElement = document.querySelector(target);
+  if (!targetElement) return;
+
+  const targetPosition = targetElement.getBoundingClientRect().top;
+  const startPosition = window.scrollY;
+  const distance = targetPosition - startPosition;
+  let startTime: number | null = null;
+
+  function animation(currentTime: number) {
+    if (startTime === null) startTime = currentTime;
+    const timeElapsed = currentTime - startTime;
+    const run = ease(timeElapsed, startPosition, distance, duration);
+    window.scrollTo(0, run);
+    if (timeElapsed < duration) requestAnimationFrame(animation);
+  }
+
+  function ease(t: number, b: number, c: number, d: number) {
+    t /= d / 2;
+    if (t < 1) return (c / 2) * t * t + b;
+    t--;
+    return (-c / 2) * (t * (t - 2) - 1) + b;
+  }
+
+  requestAnimationFrame(animation);
+}
