+++ assets/src/client/_functions/vueAppFactory.ts
@@ -1,0 +1,9 @@
+import { createApp, defineComponent } from "vue";
+import store from "@/client/_store/index";
+
+export default function vueAppFactory(
+  component: ReturnType<typeof defineComponent>,
+  target: string
+) {
+  createApp(component).use(store).mount(target);
+}
