+++ assets/src/client/global/style.scss
@@ -1,0 +1,46 @@
+@use "../../../node_modules/destyle.css/destyle.min.css";
+@use "../_styles/base";
+@use "../_styles/layout" as *;
+
+#drawer-menu {
+  position: fixed;
+  top: 0;
+  left: 0;
+  right: 0;
+  bottom: 0;
+  transform: translateX(100%);
+  z-index: 2;
+}
+
+#header {
+  position: fixed;
+  top: 0;
+  left: 0;
+  right: 0;
+  transition: top 0.2s;
+  z-index: 1;
+}
+
+#contents {
+}
+
+#bottom-icons {
+  position: fixed;
+  display: flex;
+  justify-content: end;
+  padding-bottom: 10px;
+  @include Container() {
+  }
+}
+
+#footer {
+  position: fixed;
+  bottom: 0;
+  left: 0;
+  right: 0;
+}
+
+#modal {
+  position: fixed;
+  z-index: 3;
+}
