+++ assets/src/client/pages/photo/style.scss
