+++ assets/src/client/pages/profile/style.scss
@@ -1,0 +1,2 @@
+@use "../../_styles/variables" as *;
+@use "../../_styles/layout" as *;
