+++ assets/src/client/pages/video/single.scss
