--- assets/src/global/base.scss
@@ -1,2 +1,0 @@
-@use "../../node_modules/destyle.css/destyle.min.css";
-@use "../_styles/base";
