--- assets/src/global/style.scss
@@ -1,49 +1,0 @@
-@use "../_styles/utility" as *;
-
-#drawer-menu {
-  position: fixed;
-  top: 0;
-  left: 0;
-  right: 0;
-  bottom: 0;
-  transform: translateX(100%);
-  z-index: 2;
-}
-
-#menu-icon {
-  position: fixed;
-  bottom: 60px;
-  right: 20px;
-  z-index: 3;
-}
-
-#header {
-  position: fixed;
-  top: 0;
-  left: 0;
-  right: 0;
-  transition: top 0.2s;
-  z-index: 1;
-}
-
-#contents {
-}
-
-#pagetop {
-  position: fixed;
-  bottom: 4px;
-  right: 23px;
-  z-index: 1;
-}
-
-#footer {
-  position: fixed;
-  bottom: 0;
-  left: 0;
-  right: 0;
-}
-
-#modal {
-  position: fixed;
-  z-index: 3;
-}
