--- assets/src/pages/index/style.scss
@@ -1,22 +1,0 @@
-@use "../../_styles/variables" as *;
-@use "../../_styles/layout" as *;
-
-.Index {
-  height: 200vh;
-  @include Container() {
-  }
-  &__modal01 {
-    @include Modal(300, 300, 10) {
-      background-color: $colorMain;
-      color: $colorBaseLight;
-    }
-  }
-  &__modal02 {
-    @include Modal(300, 500, 10) {
-    }
-    @include mq(tab) {
-      @include Modal(800, 300, 10) {
-      }
-    }
-  }
-}
