+++ client/mnt/src/fp3/index.njk
@@ -1,0 +1,14 @@
+---
+meta:
+  title: "FP ファイナンシャルプランナー(３級) まとめ"
+  description: "FP ファイナンシャルプランナー(３級) に関する情報をまとめています。"
+styles:
+  - pages/fp3/index.css
+scripts:
+  - pages/fp3/index.js
+---
+{% extends "base.njk" %}
+
+{% block contents %}
+  <h1>FP ファイナンシャルプランナー(３級) まとめ</h1>
+{% endblock %}
